# docs/swagger.yaml
basePath: /
definitions:
...
  handler.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: john_doe
        type: string
    required:
    - password
    - username
    type: object
  handler.LoginResponse:
    properties:
      expires_in:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
  handler.RegisterRequest:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: john@example.com
        type: string
      nickname:
        example: John Doe
        type: string
      password:
        example: "123456"
        minLength: 6
        type: string
      username:
        example: john_doe
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  model.User:
    description: 用户信息
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      nickname:
        example: John Doe
        type: string
      status:
        description: '1: 正常, 0: 禁用'
        example: 1
        type: integer
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      username:
        example: john_doe
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 响应数据
      error:
        description: 错误信息
        type: string
      message:
        description: 响应消息
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a dove server API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: dove API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录并返回 JWT token
      parameters:
      - description: 登录信息
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 认证与校验
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 用户登出，将 token 加入黑名单
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证与校验
  /auth/profile:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - 认证与校验
  /auth/register:
    post:
      consumes:
      - application/json
      description: 新用户注册
      parameters:
      - description: 注册信息
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 认证与校验
...
  /timezone/convert:
    get:
      consumes:
      - application/json
      description: 将时间从一个时区转换到另一个时区
      parameters:
      - description: 时间字符串
        example: "2023-01-01 12:00:00"
        in: query
        name: time
        required: true
        type: string
      - description: 源时区
        example: Asia/Shanghai
        in: query
        name: from_timezone
        required: true
        type: string
      - description: 目标时区
        example: America/New_York
        in: query
        name: to_timezone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 转换时区
      tags:
      - 时区
...
  /timezone/format:
    get:
      consumes:
      - application/json
      description: 将时间戳格式化为指定时区的时间字符串
      parameters:
      - description: 时间戳
        example: 1672531200
        format: int64
        in: query
        name: timestamp
        required: true
        type: integer
      - description: 时区名称
        example: Asia/Shanghai
        in: query
        name: timezone
        required: true
        type: string
      - description: 时间格式
        example: "2006-01-02 15:04:05"
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 格式化时间
      tags:
      - 时区
  /timezone/parse:
    get:
      consumes:
      - application/json
      description: 在指定时区解析时间字符串
      parameters:
      - description: 时间字符串
        example: "2023-01-01 12:00:00"
        in: query
        name: time
        required: true
        type: string
      - description: 时区名称
        example: Asia/Shanghai
        in: query
        name: timezone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 解析时间字符串
      tags:
      - 时区
  /timezone/time:
    get:
      consumes:
      - application/json
      description: 获取指定时区的当前时间
      parameters:
      - description: 时区名称
        example: Asia/Shanghai
        in: query
        name: timezone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取指定时区的时间
      tags:
      - 时区
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"